Move read detection to reader plugins
Expose read-auto detection through reader operator plugins instead of the TQL factory layer. Reader plugins now provide their detection candidates, while read_auto only coordinates probing, selection, and fallback.

Fix split_lines to cap finite reads at real line boundaries and add coverage for the CSV detector regression.
