+++ libtenzir/include/tenzir/operator_plugin.hpp
@@ -19,9 +19,12 @@
 #include "tenzir/option.hpp"
 #include "tenzir/tql2/plugin.hpp"
 
+#include <functional>
 #include <mutex>
 #include <span>
+#include <string>
 #include <string_view>
+#include <vector>
 
 namespace tenzir {
 
@@ -202,6 +205,40 @@ public:
 private:
   mutable std::once_flag desc_init_flag_;
   mutable std::shared_ptr<const Description> cached_desc_;
+};
+
+struct read_detection_input {
+  std::string_view bytes;
+  bool eof = false;
+};
+
+struct read_detection_result {
+  enum class result_state {
+    reject,
+    need_more,
+    match,
+  };
+
+  result_state state = result_state::reject;
+  uint64_t confidence = 0;
+  std::string reason = {};
+};
+
+struct read_detection_candidate {
+  std::string format_name = {};
+  std::string operator_name = {};
+  std::string pipeline = {};
+  std::vector<std::string> after = {};
+  int64_t priority = 0;
+  std::function<read_detection_result(read_detection_input)> detect = {};
+};
+
+class ReadOperatorPlugin : public virtual OperatorPlugin {
+public:
+  virtual auto read_detection_candidates() const
+    -> std::vector<read_detection_candidate> {
+    return {};
+  }
 };
 
 enum class ArgumentType { positional, named, pipeline };
@@ -1089,6 +1126,10 @@ using _::operator_plugin::Empty;
 using _::operator_plugin::OperatorPlugin;
 using _::operator_plugin::Optimization;
 using _::operator_plugin::Optimizer;
+using _::operator_plugin::read_detection_candidate;
+using _::operator_plugin::read_detection_input;
+using _::operator_plugin::read_detection_result;
+using _::operator_plugin::ReadOperatorPlugin;
 using _::operator_plugin::Spawn;
 using _::operator_plugin::SpawnFn;
 using _::operator_plugin::SpawnWith;
