+++ test/tests/operators/read_auto/csv_many_rows.stdin
@@ -1,0 +1,4 @@
+x,y
+1,2
+3,4
+5,6
