+++ test/tests/operators/read_auto/csv_many_rows.tql
@@ -1,0 +1,3 @@
+from_stdin {
+  read_auto
+}
