+++ test/tests/operators/read_auto/csv_many_rows.txt
@@ -1,0 +1,12 @@
+{
+  x: 1,
+  y: 2,
+}
+{
+  x: 3,
+  y: 4,
+}
+{
+  x: 5,
+  y: 6,
+}
